Refactor: Address review comments
