+++ tests/fixtures/tables/mod.rs
@@ -15,5 +15,6 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
+pub mod auto_sales;
 pub mod duckdb_types;
 pub mod nyc_trips;
